+++ Simulation/src/SimMaterialData.cc
@@ -16,10 +16,10 @@ void  SimMaterialData::Load(const std::string& dataDir, int verbose) {
   // first 1 lines is comment followed by the secondary e- and gamma production
   // thresholds in [MeV]
   fgets(name, sizeof(name), f);
-  fscanf(f, "%lg  %lg\n", &fElectronCut, &fGammaCut);
+  fscanf(f, "%g  %g\n", &fElectronCut, &fGammaCut);
   // one line comment followed by the 3 MSC max-step lenght paraeters
   fgets(name, sizeof(name), f);
-  fscanf(f, "%lg  %lg  %lg\n", &fMSCStepParSLow, &fMSCStepParSHigh, &fMSCStepParEcross);
+  fscanf(f, "%g  %lg  %g\n", &fMSCStepParSLow, &fMSCStepParSHigh, &fMSCStepParEcross);
   // one line comment followed by the number of materials
   fgets(name, sizeof(name), f);
   fscanf(f, "%d\n", &fNumMaterial);
@@ -36,15 +36,15 @@ void  SimMaterialData::Load(const std::string& dataDir, int verbose) {
     fgets(name, sizeof(name), f);
     fgets(name, sizeof(name), f);
     int idum;
-    double ddum;
-    fscanf(f, "%d %s %lg\n",&idum, name, &ddum);
+    float ddum;
+    fscanf(f, "%d %s %g\n",&idum, name, &ddum);
     fMaterialName[im]   = std::string(name);
     fMaterialDensity[im] = ddum;
     // 3 lines not imprtant followed by 2 numbers: the first is the Moller IMFP scaling factor
     fgets(name, sizeof(name), f);
     fgets(name, sizeof(name), f);
     fgets(name, sizeof(name), f);
-    fscanf(f, "%lg %lg\n",&(fMollerIMFPScaling[im]), &ddum);
+    fscanf(f, "%g %g\n",&(fMollerIMFPScaling[im]), &ddum);
     if (verbose > 0) {
       std::cout << "    --- Material data for: " << fMaterialName[im] << std::endl;
     }
